+++ R/qenv-get_warnings.R
@@ -28,27 +28,7 @@ setGeneric("get_warnings", function(object) {
 })
 
 setMethod("get_warnings", signature = "qenv", function(object) {
-  warnings <- lapply(object@code, "attr", "warning")
-  idx_warn <- which(sapply(warnings, function(x) !is.null(x) && !identical(x, "")))
-  if (!any(idx_warn)) {
-    return(NULL)
-  }
-  warnings <- warnings[idx_warn]
-  code <- object@code[idx_warn]
-
-  lines <- mapply(
-    function(warn, expr) {
-      sprintf("%swhen running code:\n%s", warn, expr)
-    },
-    warn = warnings,
-    expr = code
-  )
-
-  sprintf(
-    "~~~ Warnings ~~~\n\n%s\n\n~~~ Trace ~~~\n\n%s",
-    paste(lines, collapse = "\n\n"),
-    paste(get_code(object), collapse = "\n")
-  )
+  get_warn_message_util(object, "warning")
 })
 
 setMethod("get_warnings", signature = "qenv.error", function(object) {