+++ tests/testthat/test-qenv_eval_code.R
@@ -171,8 +171,6 @@ testthat::test_that(
   }
 )
 
-
-
 # comments --------------------------------------------------------------------------------------------------------
 
 testthat::test_that("comments fall into proper calls", {